nvme: move ns commands into a plugin
There are too many top-level commands cluttering the space. Move the
namespace commands into a plugin, e.g.,

  nvme delete-ns -> nvme ns delete
