+++ tests/unit/py/test_command_metadata_schema.py
@@ -168,7 +168,7 @@ class TestCommandMetadataSchema(unittest.TestCase):
     def test_expected_commands_present(self):
         """A few well-known builtin commands are emitted."""
         names = {c["name"] for c in self.data["commands"]}
-        for expected in ("list", "list-subsys", "get-ns-id", "get-log"):
+        for expected in ("list", "list-subsys", "format", "get-log"):
             self.assertIn(expected, names)
 
     def test_plugins_match_help(self):
